A big one: Support saved state in view model
0. Make ViewStateHolder and BackStack parcelable
1. Re-define view model creation mechanism to utilize saved state handles and registry owners
2. Provide convenience tools for view model creation from creator provider and view model factory creation
3. Behold the entire MVI hierarchy saved and restored upon system-initiated process death

N.B. OnRequeryFactory is restricted to the library group, and defaultArgs is private and cannot be set outside the AbstractSavedStateViewModelFactory constructor. Therefore, we cannot create a factory that selects sub-factories based on view model class and provides individual default arguments to it. Instead, we wrap view model creation function and use factory provider to pre-select view model factory from the injected aggregated map based on the reified view model type